Add filtering to Multi
