+++ packages/grafana-ui/src/components/Combobox/MultiCombobox.tsx
@@ -13,15 +13,10 @@ import { Spinner } from '../Spinner/Spinner';
 import { Text } from '../Text/Text';
 import { Tooltip } from '../Tooltip';
 
-import {
-  ComboboxOption,
-  ComboboxBaseProps,
-  AutoSizeConditionals,
-  itemToString,
-  VIRTUAL_OVERSCAN_ITEMS,
-} from './Combobox';
+import { ComboboxOption, ComboboxBaseProps, AutoSizeConditionals, VIRTUAL_OVERSCAN_ITEMS } from './Combobox';
 import { OptionListItem } from './OptionListItem';
 import { ValuePill } from './ValuePill';
+import { itemFilter, itemToString } from './filter';
 import { getComboboxStyles, MENU_OPTION_HEIGHT, MENU_OPTION_HEIGHT_DESCRIPTION } from './getComboboxStyles';
 import { getMultiComboboxStyles } from './getMultiComboboxStyles';
 import { useComboboxFloat } from './useComboboxFloat';
@@ -48,8 +43,11 @@ export const MultiCombobox = <T extends string | number>(props: MultiComboboxPro
   }, [value, options, isAsync]);
 
   const styles = useStyles2(getComboboxStyles);
+  const [inputValue, setInputValue] = useState('');
 
-  const [items, baseSetItems] = useState(isAsync ? [] : options);
+  const [baseItems, baseSetItems] = useState(isAsync ? [] : options);
+
+  const items = useMemo(() => baseItems.filter(itemFilter(inputValue)), [baseItems, inputValue]);
 
   // TODO: Improve this with async
   useEffect(() => {
@@ -72,8 +70,6 @@ export const MultiCombobox = <T extends string | number>(props: MultiComboboxPro
     (item: ComboboxOption<T>) => selectedItems.some((opt) => opt.value === item.value),
     [selectedItems]
   );
-
-  const [inputValue, setInputValue] = useState('');
 
   const { getSelectedItemProps, getDropdownProps, removeSelectedItem } = useMultipleSelection({
     selectedItems, //initally selected items,
@@ -120,7 +116,6 @@ export const MultiCombobox = <T extends string | number>(props: MultiComboboxPro
         case useCombobox.stateChangeTypes.InputBlur:
           setInputValue('');
           setIsOpen(false);
-          return changes;
         default:
           return changes;
       }
